Make Cron infocenter output more quiet

+++ releng/org.eclipse.rse.releng.infocenter/update.sh
@@ -8,21 +8,27 @@ NEED_RESTART=0
 
 #update RSE into deplopyment directory
 cd $IHOME/deploy/rse
-rm *
-wget "http://download.eclipse.org/dsdp/tm/downloads/drops/N.latest/RSE-SDK-latest.zip"
+rm *.zip
+wget -q "http://download.eclipse.org/dsdp/tm/downloads/drops/N.latest/RSE-SDK-latest.zip"
 if [ -e RSE-SDK-latest.zip ]; then
-  unzip RSE-SDK-latest.zip
-  rm -rf plugins.tmp
+  unzip -q RSE-SDK-latest.zip
+  if [ -e plugins.tmp ]; then
+    rm -rf plugins.tmp
+  fi
   mkdir plugins.tmp
   mv eclipse/plugins/*doc* plugins.tmp
   rm -rf eclipse
   NUM=`ls plugins.tmp/*.jar | wc -l`
   echo "RSE plugins.tmp: NUM=$NUM"
   if [ "$NUM" = "3" ]; then
-    rm -rf plugins
+    if [ -e plugins ]; then 
+      rm -rf plugins
+    fi
     mv plugins.tmp plugins
     NEED_RESTART=1
   fi
+else
+  echo "Error downloading RSE-SDK-latest.zip"
 fi
 
 #update Infocenter with latest deployable plug-ins
@@ -44,7 +50,7 @@ if [ "$NEED_RESTART" != "0" ]; then
   echo "Waiting for Infocenter / Apache to come up [60 seconds]"
   sleep 60
   echo "Doing fake search to force rebuilding index"
-  wget "http://localhost/help/latest/advanced/searchView.jsp?searchWord=SystemBasePlugin&maxHits=2" -O search.out.jsp -t 3 --waitretry=30
+  wget -q "http://localhost/help/latest/advanced/searchView.jsp?searchWord=SystemBasePlugin&maxHits=2" -O search.out.jsp -t 3 --waitretry=30
   echo "Done, index should be up again!"
 else
   echo "Nothing new deployed, no restart necessary."
